backend/pyproject.toml
[project]
name = "super-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.2.0",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "django-types>=0.21.0",
    "djangorestframework-stubs>=3.16.0",
]
